requirements.txt
streamlit>=1.38.0
fastapi
uvicorn[standard]
python-multipart
pydantic
opencv-python-headless
numpy
pillow
scikit-image
sqlmodel
plotly>=5.24.0
torch
torchvision
requests

